Fix type conversion warning

+++ src/app/main.cpp
@@ -140,13 +140,13 @@ int main(int argc, char* argv[])
       nthPrime(opt);
     else if (opt.RiemannR)
     {
-      long double res = RiemannR(opt.numbers[0]);
+      long double res = RiemannR((long double) opt.numbers[0]);
       int precision = (res != std::floor(res)) ? 3 : 0;
       std::cout << std::fixed << std::setprecision(precision) << res << std::endl;
     }
     else if (opt.RiemannR_inverse)
     {
-      long double res = RiemannR_inverse(opt.numbers[0]);
+      long double res = RiemannR_inverse((long double) opt.numbers[0]);
       int precision = (res != std::floor(res)) ? 3 : 0;
       std::cout << std::fixed << std::setprecision(precision) << res << std::endl;
     }
